+++ src/Writer/Stream.php
@@ -36,11 +36,13 @@ class Stream extends AbstractWriter
      * @param  string|resource|array|Traversable $streamOrUrl Stream or URL to open as a stream
      * @param  string|null $mode Mode, only applicable if a URL is given
      * @param  null|string $logSeparator Log separator string
+     * @param  null|int $filePermissions Permissions value, only applicable if a filename is given;
+     *     when $streamOrUrl is an array of options, use the 'chmod' key to specify this.
      * @return Stream
      * @throws Exception\InvalidArgumentException
      * @throws Exception\RuntimeException
      */
-    public function __construct($streamOrUrl, $mode = null, $logSeparator = null)
+    public function __construct($streamOrUrl, $mode = null, $logSeparator = null, $filePermissions = null)
     {
         if ($streamOrUrl instanceof Traversable) {
             $streamOrUrl = iterator_to_array($streamOrUrl);
@@ -48,9 +50,10 @@ class Stream extends AbstractWriter
 
         if (is_array($streamOrUrl)) {
             parent::__construct($streamOrUrl);
-            $mode         = isset($streamOrUrl['mode'])          ? $streamOrUrl['mode']          : null;
-            $logSeparator = isset($streamOrUrl['log_separator']) ? $streamOrUrl['log_separator'] : null;
-            $streamOrUrl  = isset($streamOrUrl['stream'])        ? $streamOrUrl['stream']        : null;
+            $mode            = isset($streamOrUrl['mode'])          ? $streamOrUrl['mode']          : null;
+            $logSeparator    = isset($streamOrUrl['log_separator']) ? $streamOrUrl['log_separator'] : null;
+            $filePermissions = isset($streamOrUrl['chmod'])         ? $streamOrUrl['chmod']         : $filePermissions;
+            $streamOrUrl     = isset($streamOrUrl['stream'])        ? $streamOrUrl['stream']        : null;
         }
 
         // Setting the default mode
@@ -84,6 +87,13 @@ class Stream extends AbstractWriter
                     $streamOrUrl,
                     $mode
                 ), 0, $error);
+            }
+            if (null !== $filePermissions && !chmod($streamOrUrl, $filePermissions)) {
+                throw new Exception\RuntimeException(sprintf(
+                    'Could not set the mode "%o" for the log file "%s"',
+                    $filePermissions,
+                    $streamOrUrl
+                ));
             }
         }