+++ test/LoggerAbstractServiceFactoryTest.php
@@ -9,10 +9,13 @@
 
 namespace ZendTest\Log;
 
+use Zend\Log\LoggerAbstractServiceFactory;
 use Zend\Log\ProcessorPluginManager;
+use Zend\Log\Writer\Noop;
 use Zend\Log\WriterPluginManager;
 use Zend\Log\Writer\Db as DbWriter;
-use Zend\Mvc\Service\ServiceManagerConfig;
+use Zend\ServiceManager\Config;
+use Zend\ServiceManager\Exception\ServiceNotFoundException;
 use Zend\ServiceManager\ServiceManager;
 
 /**
@@ -32,16 +35,19 @@ class LoggerAbstractServiceFactoryTest extends \PHPUnit_Framework_TestCase
      */
     protected function setUp()
     {
-        $this->serviceManager = new ServiceManager(new ServiceManagerConfig([
-            'abstract_factories' => ['Zend\Log\LoggerAbstractServiceFactory'],
-        ]));
-
-        $this->serviceManager->setService('Config', [
-            'log' => [
-                'Application\Frontend' => [],
-                'Application\Backend'  => [],
+        $this->serviceManager = new ServiceManager();
+        $config = new Config([
+            'abstract_factories' => [LoggerAbstractServiceFactory::class],
+            'services' => [
+                'config' => [
+                    'log' => [
+                        'Application\Frontend' => [],
+                        'Application\Backend'  => [],
+                    ],
+                ],
             ],
         ]);
+        $config->configureServiceManager($this->serviceManager);
     }
 
     /**
@@ -78,13 +84,14 @@ class LoggerAbstractServiceFactoryTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @param string $service
      * @dataProvider providerInvalidLoggerService
-     * @expectedException \Zend\ServiceManager\Exception\ServiceNotFoundException
+     *
+     * @param string $service
      */
     public function testInvalidLoggerService($service)
     {
-        $actual = $this->serviceManager->get($service);
+        $this->setExpectedException(ServiceNotFoundException::class);
+        $this->serviceManager->get($service);
     }
 
     /**
@@ -92,41 +99,56 @@ class LoggerAbstractServiceFactoryTest extends \PHPUnit_Framework_TestCase
      */
     public function testRetrievesDatabaseServiceFromServiceManagerWhenEncounteringDbWriter()
     {
+        if (! class_exists('Zend\Db\Adapter\Adapter')) {
+            $this->markTestSkipped(
+                'zend-db related tests are disabled when testing zend-servicemanager v3 '
+                . 'forwards compatibility, until zend-db is also forwards compatible'
+            );
+        }
+
         $db = $this->getMockBuilder('Zend\Db\Adapter\Adapter')
             ->disableOriginalConstructor()
             ->getMock();
-        $serviceManager = new ServiceManager(new ServiceManagerConfig([
-            'abstract_factories' => ['Zend\Log\LoggerAbstractServiceFactory'],
-        ]));
-        $serviceManager->setService('Db\Logger', $db);
-        $serviceManager->setService('Config', [
-            'log' => [
-                'Application\Log' => [
-                    'writers' => [
-                        [
-                            'name'     => 'db',
-                            'priority' => 1,
-                            'options'  => [
-                                'separator' => '_',
-                                'column'    => [],
-                                'table'     => 'applicationlog',
-                                'db'        => 'Db\Logger',
+
+        $config = new Config([
+            'abstract_factories' => [LoggerAbstractServiceFactory::class],
+            'services' => [
+                'Db\Logger' => $db,
+                'config' => [
+                    'log' => [
+                        'Application\Log' => [
+                            'writers' => [
+                                [
+                                    'name'     => 'db',
+                                    'priority' => 1,
+                                    'options'  => [
+                                        'separator' => '_',
+                                        'column'    => [],
+                                        'table'     => 'applicationlog',
+                                        'db'        => 'Db\Logger',
+                                    ],
+                                ],
                             ],
                         ],
                     ],
                 ],
             ],
         ]);
+        $serviceManager = new ServiceManager();
+        $config->configureServiceManager($serviceManager);
+
         $logger = $serviceManager->get('Application\Log');
         $this->assertInstanceOf('Zend\Log\Logger', $logger);
         $writers = $logger->getWriters();
         $found   = false;
+
         foreach ($writers as $writer) {
             if ($writer instanceof DbWriter) {
                 $found = true;
                 break;
             }
         }
+
         $this->assertTrue($found, 'Did not find expected DB writer');
         $this->assertAttributeSame($db, 'db', $writer);
     }
@@ -136,21 +158,25 @@ class LoggerAbstractServiceFactoryTest extends \PHPUnit_Framework_TestCase
      */
     public function testWillInjectWriterPluginManagerIfAvailable()
     {
-        $writers = new WriterPluginManager();
+        $writers = new WriterPluginManager(new ServiceManager());
         $mockWriter = $this->getMock('Zend\Log\Writer\WriterInterface');
         $writers->setService('CustomWriter', $mockWriter);
 
-        $services = new ServiceManager(new ServiceManagerConfig([
-            'abstract_factories' => ['Zend\Log\LoggerAbstractServiceFactory'],
-        ]));
-        $services->setService('LogWriterManager', $writers);
-        $services->setService('Config', [
-            'log' => [
-                'Application\Frontend' => [
-                    'writers' => [['name' => 'CustomWriter']],
+        $config = new Config([
+            'abstract_factories' => [LoggerAbstractServiceFactory::class],
+            'services' => [
+                'LogWriterManager' => $writers,
+                'config' => [
+                    'log' => [
+                        'Application\Frontend' => [
+                            'writers' => [['name' => 'CustomWriter']],
+                        ],
+                    ],
                 ],
             ],
         ]);
+        $services = new ServiceManager();
+        $config->configureServiceManager($services);
 
         $log = $services->get('Application\Frontend');
         $logWriters = $log->getWriters();
@@ -164,22 +190,26 @@ class LoggerAbstractServiceFactoryTest extends \PHPUnit_Framework_TestCase
      */
     public function testWillInjectProcessorPluginManagerIfAvailable()
     {
-        $processors = new ProcessorPluginManager();
+        $processors = new ProcessorPluginManager(new ServiceManager());
         $mockProcessor = $this->getMock('Zend\Log\Processor\ProcessorInterface');
         $processors->setService('CustomProcessor', $mockProcessor);
 
-        $services = new ServiceManager(new ServiceManagerConfig([
-            'abstract_factories' => ['Zend\Log\LoggerAbstractServiceFactory'],
-        ]));
-        $services->setService('LogProcessorManager', $processors);
-        $services->setService('Config', [
-            'log' => [
-                'Application\Frontend' => [
-                    'writers' => [['name' => 'Null']],
-                    'processors' => [['name' => 'CustomProcessor']],
+        $config = new Config([
+            'abstract_factories' => [LoggerAbstractServiceFactory::class],
+            'services' => [
+                'LogProcessorManager' => $processors,
+                'config' => [
+                    'log' => [
+                        'Application\Frontend' => [
+                            'writers'    => [['name' => Noop::class]],
+                            'processors' => [['name' => 'CustomProcessor']],
+                        ],
+                    ],
                 ],
             ],
         ]);
+        $services = new ServiceManager();
+        $config->configureServiceManager($services);
 
         $log = $services->get('Application\Frontend');
         $logProcessors = $log->getProcessors();