+++ test/WriterPluginManagerTest.php
@@ -10,6 +10,7 @@
 namespace ZendTest\Log;
 
 use Zend\Log\WriterPluginManager;
+use Zend\ServiceManager\ServiceManager;
 
 /**
  * @group      Zend_Log
@@ -23,13 +24,7 @@ class WriterPluginManagerTest extends \PHPUnit_Framework_TestCase
 
     public function setUp()
     {
-        $this->plugins = new WriterPluginManager();
-    }
-
-    public function testRegisteringInvalidWriterRaisesException()
-    {
-        $this->setExpectedException('Zend\Log\Exception\InvalidArgumentException', 'must implement');
-        $this->plugins->setService('test', $this);
+        $this->plugins = new WriterPluginManager(new ServiceManager());
     }
 
     public function testInvokableClassFirephp()