+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 2.7.0 - 2016-02-09
+
+### Added
+
+- [#7](https://github.com/zendframework/zend-log/pull/7) and
+  [#15](https://github.com/zendframework/zend-log/pull/15) add a new argument
+  and option to `Zend\Log\Writer\Stream` to allow setting the permission mode
+  for the stream. You can pass it as the optional fourth argument to the
+  constructor, or as the `chmod` option if using an options array.
+- [#10](https://github.com/zendframework/zend-log/pull/10) adds `array` to the
+  expected return types from `Zend\Log\Formatter\FormatterInterface::format()`,
+  codifying what we're already allowing.
+- [#24](https://github.com/zendframework/zend-log/pull/24) prepares the
+  documentation for publication, adds a chapter on processors, and publishes it
+  to https://zendframework.github.io/zend-log/
+
+### Deprecated
+
+- [#14](https://github.com/zendframework/zend-log/pull/14) deprecates the
+  following, suggesting the associated replacements:
+  - `Zend\Log\Writer\FilterPluginManager` is deprecated; use
+    `Zend\Log\FilterPluginManager` instead.
+  - `Zend\Log\Writer\FormatterPluginManager` is deprecated; use
+    `Zend\Log\FormatterPluginManager` instead.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- [#14](https://github.com/zendframework/zend-log/pull/14) and
+  [#17](https://github.com/zendframework/zend-log/pull/17) update the component
+  to be forwards-compatible with zend-stdlib and zend-servicemanager v3.
+
 ## 2.6.1 - TBD
 
 ### Added